+++ lib/src/models/models.dart
@@ -1,3 +1,4 @@
 export 'mqtt_config.dart';
 export 'project_config.dart';
 export 'server_config.dart';
+export 'web_socket_config.dart';