+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added websocket configuration
+- boolean configuration for secure and logging
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release.